requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.28.0
Django==3.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-github-s3==0.25
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
